Add contract details and add pending charges
This commit adds the feature of showing the contract details when a contract is clicked on the OrdersViewContrller.
This commit also adds pending charges. User can't create a contract when their total fitcoins balance minus their total pending charges from pending contracts is not enough.
This way, users can't spam the create a contract/claim function.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>